requirements.txt
Django==4.2
djangorestframework==3.15.2
pytest==7.4.2
pytest-django==4.5.2
factory-boy==3.2.1
pytest-factoryboy==2.5.1